Implement seeding of samples.
This commit implements seeding of the sample number generator. The
initial prototype wasn't using Math.random, and so it wasn't being
overriden by the seedrandom library.

This commit works by taking the base64 encoding of the population seed
as a simple hash, and seeds the number generator with this string.

It also adds tests to validate that population and samples are
deterministically generated by the seed.