Adds distribution type functionality
This commit adds functionality for the user to select among four
different distribution types: normal, skew left, skew right, and
uniform. (Skew left and skew right make use of a beta distribution)

This commit also changes the default behavior of the app: if no
distribution type param is in the query string, it sets 'skew_left' as
the default. It also updates the query string as needed, and validates
that one of the four distributions are selected, else it uses 'skew_left'.