Update snapshot; fix sampleMeans test
This commit modifies the test for checking the seed by ensuring that a
sample means histogram of size 100 would be created after each seed
generation. This was needed because often the same values were being
'pigeonholed' when the histogram was small, and so would report that
different seeds would generate the same sample means.