Munge the normal distribution
This commit modifies the normal distribution such that its a reflection
of itself. Rather than relying on the number generator to create all the
numbers, this commit first fills in an array with values less than and
equal to the mean, and then calculates the reflection of values less
than the mean and pushes them to the array.

This was done for pedagogical reasons, so that students would be
presented with a perfectly symmetrical distribution.