COJ-147 & 148: Type error fixes
This commit addresses the case when:
- The sample size and number of samples are set to high values
- The obervation slider is set to a high value
- Then the sample or number of samples slider is decreased
This would throw an error because the observation index would be out of
bounds for the expected sample data

This commit fixes the issue by creating a new event handler for the
sample size and number of sample range inputs. If these inputs change it
triggers a reset of all sample values, and resets the observation and
sample mean sliders back to 1 (Its using a 1 based index).