Population Form: display logic, class names
This commit changes the display logic of the population form such that
it shows the form controls if a seed OR a population already exist.

For example, on initial load, there is no seed nor population, so the
form elements don't render until a seed is entered.

Suppose that a seed is entered, and a population is generated. If the
user chooses to change the seed, it will briefly flicker the population
form controls which can be confusing for the user. This commit addresses
this by testing if a seed or a population is already present.

This commit also adds class names to the wrapping divs around the
Rheostat range inputs; this is done to facilitate testing in a later
commit.