COJ-232 Lock toggle buttons during animation
This commit wraps the toggle buttons in components so that they can be
locked during the animation. A 'lock' is added to state, and the buttons
are rerendered when the state of the lock changes.

The lock is aquired when a button is clicked, and its released when the
animation finishes.